# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TaskFlow API
  description: |
    API สำหรับระบบจัดการงานและทรัพยากรองค์กร (Task & Resource Management System)
    
    ## Authentication
    API ใช้ JWT Bearer Token สำหรับการยืนยันตัวตน
    ```
    Authorization: Bearer <token>
    ```
    
    ## Rate Limiting
    - Login endpoint: 5 requests/minute per IP
    - Other endpoints: ไม่จำกัด (ในเวอร์ชันปัจจุบัน)
    
    ## Roles
    - **ADMIN** - ผู้ดูแลระบบ
    - **CUSTOMER_SERVICE** - ฝ่ายบริการลูกค้า
    - **FINANCE_LEADER** - หัวหน้าฝ่ายการเงิน
    - **FINANCE** - เจ้าหน้าที่การเงิน
    - **SALES_LEADER** - หัวหน้าฝ่ายขาย
    - **SALES** - พนักงานขาย
    - **HEAD_TECH** - หัวหน้าแผนกช่าง
    - **LEADER** - หัวหน้าทีม
    - **TECH** - ช่าง
    
  version: 1.0.0
  contact:
    name: TaskFlow Support
    email: support@taskflow.com
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://taskflow.example.com/api
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token ที่ได้จาก /auth/login
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        message:
          type: string
        error:
          type: string
...
    User:
      type: object
      properties:
        id:
          type: string
        employeeId:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
        phone:
          type: string
        avatar:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        departmentId:
          type: string
        subUnitId:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    Role:
      type: string
      enum:
        - ADMIN
        - CUSTOMER_SERVICE
        - FINANCE_LEADER
        - FINANCE
        - SALES_LEADER
        - SALES
        - HEAD_TECH
        - LEADER
        - TECH
...
    CarStatus:
      type: string
      enum:
        - AVAILABLE
        - IN_USE
        - MAINTENANCE
        - OUT_OF_SERVICE
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
    RegisterRequest:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        phone:
          type: string
        departmentId:
          type: string
...
    CreateCarRequest:
      type: object
      required:
        - plateNumber
        - name
      properties:
        plateNumber:
          type: string
          minLength: 1
          maxLength: 20
        name:
          type: string
          minLength: 1
          maxLength: 100
        type:
          type: string
        brand:
          type: string
        model:
          type: string
        description:
          type: string
...
  responses:
    Unauthorized:
      description: ไม่ได้ยืนยันตัวตน
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            success: false
            error: กรุณาเข้าสู่ระบบ
    Forbidden:
      description: ไม่มีสิทธิ์เข้าถึง
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            success: false
            error: ไม่มีสิทธิ์เข้าถึง
    BadRequest:
      description: ข้อมูลไม่ถูกต้อง
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            success: false
            error: กรุณากรอกข้อมูลให้ครบถ้วน
...
    TooManyRequests:
      description: ส่งคำขอมากเกินไป
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            success: false
            error: ขออภัย คุณส่งคำขอมากเกินไป กรุณารอสักครู่
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: เข้าสู่ระบบ
      description: Login ด้วย email และ password เพื่อรับ JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: เข้าสู่ระบบสำเร็จ
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          token:
                            type: string
                          user:
                            $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: รหัสผ่านไม่ถูกต้อง
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /auth/register:
    post:
      tags: [Auth]
      summary: ลงทะเบียนผู้ใช้ใหม่
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: ลงทะเบียนสำเร็จ
        '400':
          $ref: '#/components/responses/BadRequest'
  /auth/reset-password:
    post:
      tags: [Auth]
      summary: ขอรีเซ็ตรหัสผ่าน
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: ส่งอีเมลรีเซ็ตรหัสผ่านแล้ว (ถ้าอีเมลมีในระบบ)
    patch:
      tags: [Auth]
      summary: รีเซ็ตรหัสผ่านด้วย token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token:
                  type: string
                password:
                  type: string
                  minLength: 6
      responses:
        '200':
          description: รีเซ็ตรหัสผ่านสำเร็จ
        '400':
          description: Token ไม่ถูกต้องหรือหมดอายุ
...
  /cars:
    get:
      tags: [Cars]
      summary: ดึงรายการรถ
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/CarStatus'
      responses:
        '200':
          description: รายการรถ
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Cars]
      summary: เพิ่มรถใหม่
      description: ต้องมีสิทธิ์ canManageCars (HEAD_TECH, ADMIN)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCarRequest'
      responses:
        '200':
          description: เพิ่มรถสำเร็จ
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ==================== Leaves ====================
...